+++ app/javascript/components/forms/SignupForm.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button, Stack } from 'react-bootstrap';
 import { useForm } from 'react-hook-form';
-import { FormControl } from './FormControl';
+import FormControl from './FormControl';
 import Form from './Form';
 import { signupFormConfig, signupFormFields } from '../../helpers/forms/SignupFormHelpers';
 import Spinner from '../stylings/Spinner';
