+++ app/javascript/components/stylings/buttons/ButtonLink.jsx
@@ -1,16 +1,20 @@
 import { useNavigate } from 'react-router-dom';
 import Button from 'react-bootstrap/Button';
 import React, { useCallback } from 'react';
+import PropTypes from 'prop-types';
 
 export default function ButtonLink(props) {
   const navigate = useNavigate();
-  const { to, ...rest } = props;
+  const { to } = props;
   const handleClick = useCallback(() => { navigate(to); }, [to]);
 
   return (
     <Button
-      {...rest}
       onClick={handleClick}
     />
   );
 }
+
+ButtonLink.propTypes = {
+  to: PropTypes.string.isRequired,
+};
