+++ package.json
@@ -9,6 +9,7 @@
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.8.1",
     "esbuild": "^0.14.25",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-bootstrap": "^2.2.0",
     "react-bootstrap-icons": "^1.8.1",
